restructured frontend components to fix options rendering bug
